backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

pydantic[email]==2.6.4
pydantic-settings==2.2.1

SQLAlchemy==2.0.29
psycopg2-binary==2.9.9

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# --- dev / test
pytest
httpx==0.27.0
